// WebApi/ClientApp/src/models/responses/hoa-don/IHoaDonHangHoaDacTrung.ts
export interface IHoaDonHangHoaDacTrung {
    LHHDTrung?: number;//số khung
    SKhung?: string;//số khung
    SMay?: string;//số máy
    BKSPTVChuyen?: string;//biểm kiểm soát
    TNGHang?: string;//tên người gửi hàng
    DCNGHang?: string;//địa chỉ người gửi hàng
    MSTNGHang?: string;//MST người gửi hàng
    MDDNGHang?: string;//mã định danh người gửi hàng

}